requirements.txt
boto3
requests
lxml
python-dotenv
google-generativeai
playwright      # if using playwrigh in container
pytest-playwright  # if using playwrigh in container